wdio-config: better structure for service launcher
