--- packages/wdio-firefox-profile-service/index.js
@@ -1,4 +1,0 @@
-/**
- * no service required
- */
-module.exports = function() {}
