--- packages/wdio-firefox-profile-service/launcher.js
@@ -1,2 +1,0 @@
-const FirefoxProfileLauncher = require('./build/launcher')
-module.exports = new FirefoxProfileLauncher()
