+++ packages/wdio-firefox-profile-service/src/index.js
@@ -1,0 +1,3 @@
+import FirefoxProfileLauncher from './launcher'
+
+export const launcher = FirefoxProfileLauncher
