--- packages/wdio-sauce-service/launcher.js
@@ -1,2 +1,0 @@
-var SauceService = require('./build/launcher')
-module.exports = new SauceService()
