+++ packages/wdio-selenium-standalone-service/src/index.js
@@ -1,1 +1,4 @@
+import SeleniumStandaloneLauncher from './launcher'
+
 export default class SeleniumStandaloneService {}
+export const launcher = SeleniumStandaloneLauncher
