--- packages/wdio-testingbot-service/launcher.js
@@ -1,2 +1,0 @@
-const TestingBotService = require('./build/tb-launch-service')
-module.exports = new TestingBotService()
