+++ packages/wdio-testingbot-service/src/index.js
@@ -1,0 +1,4 @@
+import TestingBotLauncher from './launcher'
+
+export default class TestingBotService {}
+export const launcher = TestingBotLauncher
